% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chipseq_TF.R
\docType{data}
\name{Chipseq_TF}
\alias{Chipseq_TF}
\title{ChIP-seq transcription factor binding sites example data}
\format{A data frame with 1000 rows and 2 variables, each variable represents a replicate of ChIP-seq read counts
 \describe{
 \item{Rep1}{ChIP-seq TF reads count for replicate one}
 \item{Rep2}{ChIP-seq TF reads count for replicate two}
 }}
\source{
Tao Yang (2016).
}
\usage{
data(Chipseq_TF)
}
\description{
A dataset containing two replicates of ChIP-seq TF data.
 The variables are as follows:
}
\references{
Evaluating the Reproducibility and Quality of High Throughput Sequencing Data with Tail Dependences
of Mixture Copula (2016). Tao Yang, Feng Yue, Qunhua Li.
}
\keyword{data}
\keyword{example}

