% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empdist.R
\name{empdist}
\alias{empdist}
\title{compute the empirical cumulative density.}
\usage{
empdist(x1, x2)
}
\arguments{
\item{x1}{a vector of data for replicate 1, i.g., sequencing counts for each genomic locus.}

\item{x2}{a vector of data for replicate 2. It must have the same length with x1.}
}
\value{
The empdist function returns a matrix that has two columns, each stores the empirical
cumulative density of one replicate.
}
\description{
compute the empirical cumulative density given two random variables,
which will be the input for the mixture copula model
}
\details{
The data was first ranked with a ties method equals to "random". The cumulative density
function (cdf) is then obtained using the ecdf function. To avoid infinity, a factor is multiplied
to the cdf. The factor is the \code{length(x1)/[length(x1)+1]}.
}
\examples{
data(Chipseq_TF)
x1 <- Chipseq_TF[,1]
x2 <- Chipseq_TF[,2]
empdist(x1, x2)
}
\references{
Q. Li, J. B. Brown, H. Huang and P. J. Bickel. (2011) Measuring reproducibility of high-throughput
experiments. Annals of Applied Statistics, Vol. 5, No. 3, 1752-1779.
}

