% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqcush.R
\name{seqcush}
\alias{seqcush}
\title{Add cushion for sequencing data.}
\usage{
seqcush(dat, cush)
}
\arguments{
\item{dat}{a matrix that has two columns, each of them represent a replicate of data,
i.g., sequencing counts for genomic loci.}

\item{cush}{an integer, the differece in sequencing counts that are deemed the same.
For example, if a cush of 2 is chosen, the read count of 4 at a loci of replicate 1 and
the read count of 6 at the same loci of replicate 2 are deemed the same. The 6 in replicate 2
is corrected to 4.}
}
\value{
This function returns a matrix of two colunms of data, i.g., seqencing counts. Each
column is a vector a counts after correction to eliminate the discordance between two
replicates in ranks.
}
\description{
Add a cushion for sequencing data which reduces the discrepancy of read counts that
has no biological meaning but greatly increases the discordance in ranks of the data.
}
\examples{
data(Chipseq_TF)
cush <- 2
seqcush(Chipseq_TF, cush)
}

