% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TailD.R
\name{TailD}
\alias{TailD}
\title{extracting proportion, upper tail dependence and lower tail dependence}
\usage{
TailD(object)
}
\arguments{
\item{object}{the output from the function TDR.est}
}
\value{
\itemize{
 \item {pi} {the proportion parameter that indicates how much of mixture model is contributed by Clayton copula.}
 \item {UT} {the upper tail dependence coefficient.}
 \item {LT} {the lower tail dependence coefficient.}
}
}
\description{
function to extract proportion, upper tail dependence and lower tail dependence
}
\examples{
data(Chipseq_TF)
x1 <- Chipseq_TF[,1]
x2 <- Chipseq_TF[,2]
U=empdist(x1, x2)
u <- U[,1]
v <- U[,2]
test <- TDR.est(u, v)
TailD(test)
}
\references{
Evaluating the Reproducibility and Quality of High Throughput Sequencing Data with Tail Dependences
of Mixture Copula (2016). Tao Yang, Feng Yue, Qunhua Li.
}

