% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdCreate.R
\name{tdCreate}
\alias{tdCreate}
\title{tdCreate}
\usage{
tdCreate(data = NULL, table = NULL, upload = TRUE, colType = NULL,
  pi = NULL, batchSize = 2500, verbose = TRUE, ...)
}
\arguments{
\item{data}{\code{link{data.frame}} containing data to upload. This must be in
the same column order as the Teradata table.}

\item{table}{Name of the Teradata table to upload data to. If \code{NULL}, then the name of the
\code{data.frame} provided will be used as the Teradata table name, with periods replaced by
underscores.}

\item{upload}{logical. If \code{TRUE}, then the function will call \code{tdUpload} to
upload the data after creating the table.}

\item{colType}{Vector of column types for Teradata table. If NULL, will be inferred following
approach listed in details section.}

\item{pi}{Primary index for Teradata table. If NULL, will take the first column with all unique
values as the primary index. If none exists, will take the first column in \code{data}.}

\item{batchSize}{Number of rows to upload simultaneously. Only used if \code{upload=TRUE}.}

\item{verbose}{logical. If \code{TRUE}, then print message after each batch is uploaded.}

\item{...}{Optional connection settings.}
}
\value{
Returns an invisible objecting containing the count of the number of rows uploaded.
}
\description{
Creates a table in Teradata and (optionally) uploads data. Can take a JDBC
connection object (\code{conn}) if provided. If no JDBC connection is provided,
then a connection is attempted using the \code{user}, and \code{password}
provided. If none is provided, then tries to locate a connection object
(\code{conn}) in the global environment.

If a connection profile (e.g. username, password, etc.) is provided, then
an attempt is made to connect to Teradata. Once the query is run, the
connection is then closed. If a connection object (\code{conn}) is provided
to the function (or one is found globally), then the connection remains
open.
}
\details{
If not provided, the columns will be defined according to the class of each column in
the provided \code{data.frame}.
\itemize{
	\item{numeric}{Integer if maximum value is under 2,000,000. Otherwise, bigint.}
	\item{date}{Date}
	\item{POSIXct or POSIXt}{Timestamp(0)}
	\item{All others}{varchar with length equal to the longest character length}
'}

Please note that in order to keep the required Java memory down, data is uploaded in batches
at a time. This can be configured to be as many (or little observations) as desired.
}
\seealso{
\code{\link{tdConn}} for connection, \code{\link{tdUpload}} for table upload,
\code{\link{td}} for general queries, \code{\link{tdCpu}} for CPU
usage.
}

