% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdDisk.R
\name{tdDisk}
\alias{tdDisk}
\title{tdDisk}
\usage{
tdDisk(user = "user", ...)
}
\arguments{
\item{user}{Username to grab CPU use from. Defaults to the \code{username}
used in the Teradata connection.}

\item{...}{Optional connection settings.}
}
\value{
A \code{\link{data.frame}} object is returned with all of the Teradata
query information of the specified date.
}
\description{
Queries Teradata for disk space used. This code is specifically
designed for connectivity to Teradata servers using OSX at Apple using JDBC
drivers and should be updated if connected to other sources. Can take a JDBC
connection object (\code{conn}) if provided. If no JDBC connection is provided,
then a connection is attempted using the \code{user}, and \code{password}
provided. If none is provided, then tries to locate a connection object
(\code{conn}) in the global environment.

If a connection profile (e.g. username, password, etc.) is provided, then
an attempt is made to connect to Teradata. Once the query is run, the
connection is then closed. If a connection object (\code{conn}) is provided
to the function (or one is found globally), then the connection remains
open.
}
\examples{
## NOT RUN ##
## Connect to default data warehouse and data base
# tdDisk(<username>, <password>)

## Runs query using a separately established connection
# conn = tdConn(<username>, <password>)
# tdDisk(conn=conn)

## Uses same connection, but allows code to find globally
# tdDisk()

}
\seealso{
\code{\link{tdConn}} for connection, \code{\link{tdCpu}} for CPU usage,
\code{\link{tdSpool}} for spool usage, and \code{\link{td}} for general 
queries
}

