% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdExists.R
\name{tdExists}
\alias{tdExists}
\title{tdExists}
\usage{
tdExists(table = NULL, ...)
}
\arguments{
\item{table}{String vector of name of table to get column names from.}

\item{...}{Optional connection settings.}
}
\value{
Returns a boolean indicator of either \code{TRUE} if table exists or 
\code{FALSE} if table does not.
}
\description{
Determines if Teradata table exists. This code is specifically 
designed for connectivity to Teradata servers using OSX at Apple using JDBC 
drivers and should be updated if connected to other sources. Can take a JDBC 
connection object (\code{conn}) if provided. If no JDBC connection is provided, 
then a connection is attempted using the \code{user}, and \code{password} 
provided. If none is provided, then tries to locate a connection object 
(\code{conn}) in the global environment.

If a connection profile (e.g. username, password, etc.) is provided, then
an attempt is made to connect to Teradata. Once the query is run, the
connection is then closed. If a connection object (\code{conn}) is provided
to the function (or one is found globally), then the connection remains
open.
}
\details{
This function looks in the \code{DBC.COLUMNS} table to determine 
whether the Teradata table exists.
}
\examples{
## NOT RUN (will also result in errors due to user restrictions) ##
## Runs a quick query based on connection profile
# tdExists("ICDB_PERSON", username=<username>, password=<password>, db="GCA")

## Runs query using a separately established connection
# conn = tdConn(<username>, <password>, db="GCA")
# tdExists("ICDB_PERSON", conn=conn)

## Uses same connection, but allows code to find globally.
# tdExists("ICDB_PERSON")

}
\seealso{
\code{\link{tdConn}} for connection, \code{\link{tdDisk}} for disk usage,
\code{\link{tdSpool}} for spool usage, \code{\link{tdCpu}} for CPU 
usage, and \code{\link{tdJoin}} for joining tables.
}

