% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdMakeSQL.R
\name{tdMakeSQL}
\alias{tdMakeSQL}
\title{tdMakeSQL}
\usage{
tdMakeSQL(file = NULL, outfile = NULL, ...)
}
\arguments{
\item{file}{File to submit to Teradata.}

\item{outfile}{Optional file name to output code to. If supplied, then code 
will be printed in file rather than console.}

\item{...}{Optional arguments to replace pointers.}
}
\value{
A string vector is returned of the SQL code.
}
\description{
Takes a SQL file and replaces pointers with user declared values. 
Can take a JDBC connection object (\code{conn}) if provided. If no 
JDBC connection is provided, then a connection is attempted using the 
\code{user}, and \code{password} provided. If none is provided, then tries 
to locate a connection object (\code{conn}) in the global environment.

If a connection profile (e.g. username, password, etc.) is provided, then
an attempt is made to connect to Teradata. Once the query is run, the
connection is then closed. If a connection object (\code{conn}) is provided
to the function (or one is found globally), then the connection remains
open.
}
\details{
\emph{Warning:} This function reads in all lines and parses commands 
using "\code{;}". Thus, commands should be separated using that character. 
If a literal ";" is desired within the code, an escape character of "\" 
should precede it, e.g. \code{where column="\;"}. Pointers should be saved
in the file using a "&" prefix. An escape character of "\" can also be used
for this if a literal "&" is desired within the code.
}
\examples{
## NOT RUN ##
## Simply returns the code from the file
# tdMakeSQL("file.sql")

## Returns code from file, replacing "&ABTExt" with "Test"
# tdMakeSQL("file.sql", ABTExt="Test")

## Returns code from file, replacing "&ABTExt" with "Test" and "&baseExt" with "ing"
# tdMakeSQL("file.sql", ABTExt="Test", baseExt="ing")

}
\seealso{
\code{\link{tdConn}} for connection, \code{\link{tdHead}} for top observations,
\code{\link{tdSpool}} for spool usage, and \code{\link{td}} for general 
queries
}

