% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDR.R
\docType{package}
\name{TDR-package}
\alias{TDR}
\alias{TDR-package}
\title{Tail dependence as reproducibility measure.}
\description{
Estimate the reproducibility of sequencing-based data using tail dependence of mixture copula model.
}
\details{
The major functions you're likely to need from \pkg{TDR} is \code{\link{TDR.est}} and \code{\link{TailD}}
}

