% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDR.est.R
\name{TDR.est}
\alias{TDR.est}
\title{estimate the paramters of the mixture copula model.}
\usage{
TDR.est(x1, x2, p = 0.25, alpha = 1.5, beta = 1.5, it = 100,
  eps = 0.001)
}
\arguments{
\item{x1}{a vector containing numeric values (i.g., sequencing counts of genomic loci) for one replicate data.}

\item{x2}{a vector containing numeric values (i.g., sequencing counts of genomic loci) for the other replicate data.}

\item{p}{initialization of proportion paramter, any value between 0 and 1 (not include).}

\item{alpha}{initialization of Gumbel copula paramter, any value between 0 and 1 (not include).}

\item{beta}{initialization of Clayton copula paramter, any value greater than 0 (not include).}

\item{it}{the maximum iterations. Usually 100 is enough to let estimation converge.}

\item{eps}{the stopping criterion. The default is 0.001, meaning if any of the paramter changes less than 0.001,
the estimation algorithm will stop.}
}
\value{
\itemize{
  \item{trace.para} {the trace of the paramters for each iteration.}
  \item{trace.loglik} {the trace of the log likelihood for each iteration.}
  \item{membership} {a vector of binary data inidicating which data points are mapped to Clayton part (1) and
                     which are mapped to Gumbel copula (0).}
 }
}
\description{
function to estimate the paramters of the mixture copula model (i.e., mixture of Gumbel and Clayton)
given the initialization parameters, maximum iteration and the stopping criterion.
}
\details{
The expectation maximization (EM) algorithm is adopted to estimate the paramters for the mixture copula model. The mixture copula model is:
\deqn{C(u, v | \pi, \alpha, \beta) = \pi * C_{clayton}(u, v | \beta) + (1-\pi) * C_{gumbel}(u, v | \alpha)}
where pi is the proportion paramters indicating how much of the mixture copula is contributed from Clayton copula, alpha is the single paramter for Gumbel copula, and the beta is the single paramter for Clayton copula.
}
\examples{
data(Chipseq_TF)
x1 <- Chipseq_TF[,1]
x2 <- Chipseq_TF[,2]
TDR.est(x1, x2)
}
\references{
Evaluating the Reproducibility and Quality of High Throughput Sequencing Data with Tail Dependences
of Mixture Copula (2016). Tao Yang, Feng Yue, Qunhua Li.
}

