% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td.R
\name{td}
\alias{td}
\title{td}
\usage{
td(query = "", ...)
}
\arguments{
\item{query}{Query string to send to Teradata.}

\item{...}{Optional connection settings.}
}
\value{
If no data is returned from query, then an \code{\link{invisible}}
object is returned. Otherwise, a \code{\link{data.frame}} object with all
data queried will be returned.
}
\description{
Sends queries to Teradata. Can take a JDBC connection
object (\code{conn}) if provided. If no JDBC connection is provided, then
a connection is attempted using the \code{user}, and \code{password} provided.
If none is provided, then tries to locate a connection object (\code{conn})
in the global environment.

If a connection profile (e.g. username, password, etc.) is provided, then
an attempt is made to connect to Teradata. Once the query is run, the
connection is then closed. If a connection object (\code{conn}) is provided
to the function (or one is found globally), then the connection remains
open.
}
\details{
Uses the v15.10.00.33 release (12 Jan 2016) tdgssconfig.jar and
terajdbc4.jar drivers.

A hidden option allows the user to print the SQL code being generated directly
to the console without submitting the query to Teradata. Simply specify
\code{.tdCat=TRUE} in the global environment before running the code.
}
\examples{
## NOT RUN ##
## Runs a quick query based on connection profile
# td("select count(*) from ICDB_PERSON", username=<username>, password=<password>, db="GCA")

## Runs query using a separately established connection
# conn = tdConn(<username>, <password>, db="GCA")
# td("select count(*) from ICDB_PERSON", conn=conn)

## Uses same connection, but allows code to find globally
# td("select count(*) from ICDB_PERSON")

}
\seealso{
\code{\link{tdConn}} for connection, \code{\link{tdDisk}} for disk usage,
\code{\link{tdSpool}} for spool usage, \code{\link{tdCpu}} for CPU
usage, and \code{\link{tdJoin}} for joining tables.
}

