% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdConn.R
\name{tdConn}
\alias{tdConn}
\title{tdConn}
\usage{
tdConn(username = NULL, password = NULL,
  addr = "jdbc:teradata://megadew.corp.apple.com/charset=utf8",
  db = ifelse(addr == "jdbc:teradata://megadew.corp.apple.com/charset=utf8",
  "CDM_Special", ""), classPath = NULL, conn = NULL)
}
\arguments{
\item{username}{Connection user name.}

\item{password}{Connection password.}

\item{addr}{String containing address of database to connect to. By default, is
\emph{jdbc:teradata://megadew.corp.apple.com/charset=utf8}.}

\item{db}{Name of database to connect to. By default, is \emph{CDM_Special}.}

\item{classPath}{The location of the JDBC drivers. By default, will use the drivers included
in the package.}

\item{conn}{\code{DBIConnection} object with established connection to the RDMBS.
Only used internally to check connection and establish a connection if none exists.}
}
\value{
A \code{RJDBC} connection object is returned.
}
\description{
Checks for a connection to the Teradata server. If none exists,
tries to establish one. This code is specifically designed for connectivity
to Teradata servers using OSX at Apple using JDBC drivers and should be updated if
connected to other sources. If no JDBC connection is provided (\code{tdConn}),
then a connection is attempted using the \code{user}, and \code{password} provided.
}
\details{
Uses the v15.10.00.33 release (12 Jan 2016) tdgssconfig.jar and
terajdbc4.jar drivers.
}
\examples{
## NOT RUN ##
## Connect to default data warehouse and data base
# conn = tdConn(<username>, <password>)

## Connect to data warehouse using different data base
# conn = tdConn(<username>, <password>, db='ADM_AMR')

## Connect to different data warehouse than default
# conn = tdConn(<username>, <password>, addr="jdbc:teradata://redwood.corp.apple.com")

}

