% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdDim.R
\name{tdDim}
\alias{tdDim}
\title{tdDim}
\usage{
tdDim(tables = NULL, where = "", ...)
}
\arguments{
\item{tables}{String vector of names of tables to get table dimensions from.}

\item{where}{String statement to subset table with.}

\item{...}{Optional connection settings.}
}
\value{
Returns a \code{\link{data.frame}} of the the Teradata table dimensions.
}
\description{
Gets the dimensions in Teradata tables. This code is specifically 
designed for connectivity to Teradata servers using OSX at Apple using JDBC 
drivers and should be updated if connected to other sources. Can take a JDBC 
connection object (\code{conn}) if provided. If no JDBC connection is provided, 
then a connection is attempted using the \code{user}, and \code{password} 
provided. If none is provided, then tries to locate a connection object 
(\code{conn}) in the global environment.

If a connection profile (e.g. username, password, etc.) is provided, then
an attempt is made to connect to Teradata. Once the query is run, the
connection is then closed. If a connection object (\code{conn}) is provided
to the function (or one is found globally), then the connection remains
open.
}
\examples{
## NOT RUN (will also result in errors due to user restrictions) ##
## Runs a quick query based on connection profile
# tdDim("ICDB_PERSON", username=<username>, password=<password>, db="GCA")

## Runs query using a separately established connection
# conn = tdConn(<username>, <password>, db="GCA")
# tdDim("ICDB_PERSON", conn=conn)

## Uses same connection, but allows code to find globally. Also used for multiple tables.
# tdDim(c("ICDB_PERSON", "ICDB_PERSON_X"))

}
\seealso{
\code{\link{tdConn}} for connection, \code{\link{tdNames}} for table names,
\code{\link{td}} for general queries, \code{\link{tdCpu}} for CPU 
usage, and \code{\link{tdJoin}} for joining tables.
}

